refactor(transformer): extract remark elm attr parse logic
