+++ packages/preset-dumi/src/transformer/remark/code.ts
@@ -3,15 +3,11 @@ import is from 'hast-util-is-element';
 import has from 'hast-util-has-property';
 import visit from 'unist-util-visit';
 import { parseText } from 'sylvanas';
+import { parseElmAttrToProps } from './utils';
 import { getModuleResolvePath } from '../../utils/moduleResolver';
 import { saveFileOnDepChange } from '../../utils/watcher';
 import transformer from '..';
 import { IDumiUnifiedTransformer, IDumiElmNode } from '.';
-
-const ATTR_MAPPING = {
-  hideactions: 'hideActions',
-  defaultshowcode: 'defaultShowCode',
-};
 
 /**
  * remark plugin for parse code tag to external demo
@@ -30,32 +26,7 @@ export default function code(): IDumiUnifiedTransformer {
 
         // read external demo content and convert node to demo node
         const result = transformer.code(fs.readFileSync(absPath).toString());
-
-        // restore camelCase attrs, because hast-util-raw will transform camlCase to lowercase
-        Object.entries(ATTR_MAPPING).forEach(([mark, attr]) => {
-          if (attrs[mark] !== undefined) {
-            attrs[attr] = attrs[mark];
-            delete attrs[mark];
-          }
-        });
-
-        // convert empty string to boolean
-        Object.keys(attrs).forEach(attr => {
-          if (attrs[attr] === '') {
-            attrs[attr] = true;
-          }
-        });
-
-        // try to parse JSON field value
-        Object.keys(attrs).forEach(attr => {
-          if (/^(\[|{)[^]*(]|})$/.test(attrs[attr])) {
-            try {
-              attrs[attr] = JSON.parse(attrs[attr]);
-            } catch (err) {
-              /* nothing */
-            }
-          }
-        });
+        const parsedAttrs = parseElmAttrToProps(attrs);
 
         // replace original node
         parent.children.splice(i, 1, {
@@ -73,7 +44,7 @@ export default function code(): IDumiUnifiedTransformer {
                 : { jsx: result.content },
             filePath: absPath,
             meta: {
-              ...attrs,
+              ...parsedAttrs,
               ...result.meta,
             },
           },
